style

+++ src/gui/gui.h
@@ -76,7 +76,7 @@ struct gui_nvram_t {
 uint_fast8_t hamradio_get_multilinemenu_block_groups(menu_names_t * vals);
 uint_fast8_t hamradio_get_multilinemenu_block_params(menu_names_t * vals, uint_fast8_t index, uint_fast8_t max_count);
 void hamradio_get_multilinemenu_block_vals(menu_names_t * vals, uint_fast8_t index, uint_fast8_t cnt);
-const char * hamradio_gui_edit_menu_item(uint_fast8_t index, int_fast8_t rotate);
+const char * hamradio_gui_edit_menu_item(uint_fast8_t index, int_least16_t rotate);
 void hamradio_clean_memory_cells(uint_fast8_t i);
 void hamradio_save_memory_cells(uint_fast8_t i);
 uint_fast32_t hamradio_load_memory_cells(uint_fast8_t cell, uint_fast8_t set);
